feat(auth): Implement custom user model TesterUser and reorganize models structure
